eth,rpc: move RPCBlock,RPCTransaction from ethapi into rpc
